[AutoDiff] NFC: Move some 'IndexSubset' method impls to their own file. (#27627)
- Move some 'IndexSubset' method implemenetations from AutoDiff.cpp to a new file called IndexSubset.cpp.
- Fix some indentation errors caused by the earlier renaming of `AutoDiffIndexSubset`.
